requirements.txt
annotated-types==0.6.0
diskcache==5.6.3
llama_cpp_python==0.2.11
nodeenv==1.8.0
numpy==1.26.1
pydantic==2.4.2
pydantic_core==2.10.1
pyright==1.1.332
typing_extensions==4.8.0
